core: add `init_examples` feature
This is mainly for development, allows for adding KMAIN_INIT_PARTIALS
that do runtime tests of various things.

This commit moves the previously reverted "write to the console using
the filesystem handlers" test into `slos::filesystem` and adds that to
the KMAIN_INIT_PARTIALS list, gated on this feature being present.