core: add `build.rs` to generate `SLOS_FEATURES` env var
This makes `env!("SLOS_FEATURES")` return a space-separated list of
the enabled Cargo features for the slos core crate.

This is done using some `cargo_toml` magic, checking that the features
passed by Cargo to the build script in the `CARGO_FEATURE_x` env vars
are actually present in the slos Cargo manifest. Doing it this way
means that the feature name in the returned list is always consistent
with the actual name of the feature, because Cargo mangles the feature
name when putting it into the environment variable.

Yes, this is why we switched to the v2 resolver, why do you ask?